Modify library loading logic to be more generic
Instead of looking for "Ubuntu" in the filename of the library to load,
iterate through a list of possible libs. If one fails, move on to the
next. This way we don't have to have any special parsing logic for the
os version, we just keep trying libs until something works.